oops
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@19905 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ e17/libs/etk/src/lib/etk_progress_bar.h
@@ -27,7 +27,7 @@ struct _Etk_Progress_Bar
    Etk_Widget *label;   
    
    unsigned char activity_dir : 1;
-   int pulse_step;
+   double pulse_step;
 };
 
 Etk_Type *etk_progress_bar_type_get();
